requirements.txt
asgiref==3.4.1
autopep8==1.6.0
cffi==1.15.0
cryptography==37.0.2
Django==3.2.13
django-cors-headers==3.10.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
Pillow==8.4.0
pkg_resources==0.0.0
pycodestyle==2.8.0
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
